requirements.txt
PyGithub
python-frontmatter
python-dotenv
# AI功能依赖
google-generativeai>=0.3.0
